[296] Move Subsction list edditing to edit profile route
