+++ 1-src/2-services/10-utilities/logging/logEntryModel.mts
@@ -1,7 +1,6 @@
 import * as log from './log.mjs';
 import { LogListItem, LogType } from '../../../0-assets/field-sync/api-type-sync/utility-types.mjs';
-import { ENVIRONMENT_TYPE } from '../../../0-assets/field-sync/input-config-sync/inputField.mjs';
-import { getEnvironment, stringifyErrorMessage } from '../utilities.mjs';
+import { stringifyErrorMessage } from '../utilities.mjs';
 import { LOG_SIMILAR_TIME_RANGE, LOG_SOURCE, PRINT_LOGS_TO_CONSOLE } from './log-types.mjs';
 import { AthenaFieldSchema } from '../athena.mjs';
